Lesson 69 (commit 14) save changes where fix bugs

+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/DTO/ProductsDTO.java
@@ -8,14 +8,15 @@ import lombok.*;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
 public class ProductsDTO {
+
     private Integer id;
     private String name;
     private String image;
     private Integer quantity;
     private String description;
-    private Integer price;
+    private Float price;
 
-    public static ProductsDTO from(Products products){
+    public static ProductsDTO from(Products products) {
         return builder()
                 .id(products.getId())
                 .name(products.getName())
