+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/Entity/Basket.java
@@ -14,26 +14,27 @@ import javax.validation.constraints.Positive;
 @Entity
 @Table(name = "baskets")
 public class Basket {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
     @NotEmpty
-    @ManyToOne (fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "products_id")
     private Products product;
 
     @NotEmpty
-    @ManyToOne (fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "customers_id")
     private User customer;
 
-    @Column( length = 255)
+    @Column(length = 255)
     private Integer quantity;
 
     @NotNull
     @NotEmpty
     @Positive
-    private Integer price;
+    private Float price;
 
 }
