+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/Service/ProductsService.java
@@ -4,7 +4,6 @@ import com.example.java_group_11_online_store_ayday_mirbekkyzy.DTO.ProductsDTO;
 import com.example.java_group_11_online_store_ayday_mirbekkyzy.Entity.Products;
 import com.example.java_group_11_online_store_ayday_mirbekkyzy.Repository.ProductsRepository;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -21,17 +20,17 @@ public class ProductsService {
         return products.map(ProductsDTO::from);
     }
 
-    public Page<ProductsDTO> searchProducts(String keyword,Pageable pageable){
-        if (keyword!=null){
-        Page<Products> products = productsRepository.search(keyword, pageable);
-        return products.map(ProductsDTO::from);}
-        else {
+    public Page<ProductsDTO> searchProducts(String keyword, Pageable pageable) {
+        if (keyword != null) {
+            Page<Products> products = productsRepository.search(keyword, pageable);
+            return products.map(ProductsDTO::from);
+        } else {
             return productsRepository.findAll(pageable).map(ProductsDTO::from);
         }
     }
 
-    public Page<ProductsDTO> priceSearch(Integer more,Integer less,Pageable pageable){
-        Page<Products> products = productsRepository.findProductsByPriceGreaterThanAndPriceIsLessThan(more,less,pageable);
+    public Page<ProductsDTO> priceSearch(Float more, Float less, Pageable pageable) {
+        Page<Products> products = productsRepository.findProductsByPriceGreaterThanAndPriceIsLessThan(more, less, pageable);
         return products.map(ProductsDTO::from);
     }
 }
